pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "systematic_weakness_detection"
version = "0.1.0"
description = "Code for TMLR paper"
authors = [{name = "Sujan Sai Gannamaneni", email = "sujan.sai.gannamaneni@iais.fraunhofer.de"}]
readme = "README.md"
requires-python = ">=3.8,<3.11"

dependencies = [
    # Core ML/CV libraries
    "torch>=2.0.0",
    "torchvision>=0.15.0", 
    "torchaudio>=2.0.0",
    "numpy>=1.24.0,<1.27.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    
    # Data science
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    
    # APIs and utilities
    "openai>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    
    # YOLO/Object detection
    "ultralytics>=8.0.0",

    # Computer Vision model libraries
    "timm==0.9.9",  # PyTorch Image Models
    
    # Other utilities
    "ftfy>=6.0.0",  # text cleaning for CLIP
    "regex>=2023.0.0",
]

[project.optional-dependencies]
dev = [
    # Jupyter and development
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
    "ipykernel>=6.0.0",
    "ipywidgets>=8.0.0",
    
    # Code quality
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    
    # Additional utilities
    "pyarrow>=10.0.0",  # for faster pandas operations
]

gpu = [
    # GPU-specific packages (if needed)
    "nvidia-ml-py3>=7.0.0",
]

all = [
    "systematic_weakness_detection[dev,gpu]"
]

[project.urls]
Homepage = "https://github.com/yourusername/your-experiment-name"
Repository = "https://github.com/yourusername/your-experiment-name"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"